refactor: apply useLegacyReactive to vue imready
